feat(machine): start 기능 구현
