+++ src/main/java/vendingmachine/machine/MachineController.java
@@ -14,27 +14,15 @@ import static vendingmachine.constant.Constant.PRODUCT_SPLITTER;
 public class MachineController {
     private final CoinController coinController = new CoinController();
     private final ProductController productController = new ProductController();
+    private final MachineService machineService = new MachineService();
 
-    public void makeCoins(Machine machine,int money) {
-        Map<Coin, Integer> coins = machine.getCoins();
-        while( money > 0 ){
-            Coin coin = coinController.pickPossibleRandomCoin(money);
-            coins.put(coin ,coins.get(coin)+1); //선택된 코인 1개 증가
-            money-=coin.getAmount();
-        }
+    public void start(Machine machine,int money,String productList){
+        machineService.makeCoins(machine,money);
+        machineService.makeProductList(machine,productList);
     }
 
-    public void makeProductList(Machine machine,String infoList){
-        List<Product> productList = machine.getProductList();
-
-        Arrays.stream(infoList.split(PRODUCT_SPLITTER))
-                .forEach(info->{
-                    productList.add(productController.makeProduct(info));
-                });
+    public void run(Machine machine,int money){
+        machineService.insertMoney(machine,money);
     }
-
-
-
-
 
 }
