+++ src/main/java/vendingmachine/product/ProductController.java
@@ -4,8 +4,5 @@ import vendingmachine.exception.ProductValidator;
 
 public class ProductController {
     private final ProductService productService = new ProductService();
-
-    public Product makeProduct(String product){
-        return productService.makeProduct(product);
-    }
+    
 }
